+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 - Add the command-line conversion tool convert-enmap (#9)
 - Include band center wavelengths in data and/or metadata (#6)
+- Include quicklook data (#7)
 
 ## Changes in 0.0.2