+++ CHANGES.md
@@ -1,4 +1,4 @@
-## Changes in 0.0.3 (under development)
+## Changes in 0.0.3
 
 - Add the command-line conversion tool convert-enmap (#9)
 - Added support for backend methods `open_datatree` and `open_groups_as_dict` (#2)